- integrated osgGaming lib

+++ trunk/osgGaming/include/osgGaming/GameSettings.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <osg/Referenced>
-#include <osg/Vec2d>
+#include <osg/Vec2i>
 
 namespace osgGaming
 {
@@ -11,10 +11,10 @@ namespace osgGaming
 		GameSettings();
 
 		void setFullscreenEnabled(bool enabled);
-		void setWindowResolution(osg::Vec2d resolution);
+		void setWindowResolution(osg::Vec2i resolution);
 
 		bool getFullscreenEnabled();
-		osg::Vec2d getWindowResolution();
+		osg::Vec2i getWindowResolution();
 
 
 		void load();
@@ -22,7 +22,7 @@ namespace osgGaming
 
 	private:
 		bool _fullscreenEnabled;
-		osg::Vec2d _windowResolution;
+		osg::Vec2i _windowResolution;
 
 		bool _isLoaded;
 	};
